renaming input
