+++ src/components/index.ts
@@ -4,6 +4,6 @@ export { RadioButtonGroup } from "./RadioButtonGroup";
 export { Heading1, Heading2, Title, Subheading, Link, ButtonTypography, Caption1, Caption2, Menu, Placeholder } from "./Typography";
 export { Toggle } from "./Toggle";
 export { Dropdown } from "./Dropdown";
-export { InputArea } from "./InputArea";
+export { Input } from "./Input";
 export { default as ThemeProvider } from "./Theme";
 export { ThemeContext } from "./Theme";
